+++ src/DeviceInterface/AppleDeviceInterface.js
@@ -1,0 +1,142 @@
+const InterfacePrototype = require('./InterfacePrototype.js')
+const {wkSend, wkSendAndWait} = require('../appleDeviceUtils/appleDeviceUtils')
+const {
+    isApp, notifyWebApp,
+    isDDGDomain,
+    formatDuckAddress
+} = require('../autofill-utils')
+const {scanForInputs, forms} = require('../scanForInputs.js')
+
+class AppleDeviceInterface extends InterfacePrototype {
+    async setupAutofill ({shouldLog} = {shouldLog: false}) {
+        if (isDDGDomain()) {
+            // Tell the web app whether we're in the app
+            notifyWebApp({isApp})
+        }
+
+        if (isApp) {
+            await this.getAutofillInitData()
+        }
+
+        const signedIn = await this._checkDeviceSignedIn()
+        if (signedIn) {
+            if (isApp) {
+                await this.getAddresses()
+            }
+            notifyWebApp({ deviceSignedIn: {value: true, shouldLog} })
+            forms.forEach(form => form.redecorateAllInputs())
+        } else {
+            this.trySigningIn()
+        }
+
+        scanForInputs(this)
+    }
+
+    async getAddresses () {
+        if (!isApp) return this.getAlias()
+
+        const {addresses} = await wkSendAndWait('emailHandlerGetAddresses')
+        this.storeLocalAddresses(addresses)
+        return addresses
+    }
+
+    async refreshAlias () {
+        await wkSendAndWait('emailHandlerRefreshAlias')
+        // On macOS we also update the addresses stored locally
+        if (isApp) this.getAddresses()
+    }
+
+    async _checkDeviceSignedIn () {
+        const {isAppSignedIn} = await wkSendAndWait('emailHandlerCheckAppSignedInStatus')
+        this.isDeviceSignedIn = () => !!isAppSignedIn
+        return !!isAppSignedIn
+    }
+
+    storeUserData ({addUserData: {token, userName, cohort}}) {
+        return wkSend('emailHandlerStoreToken', { token, username: userName, cohort })
+    }
+
+    /**
+     * PM endpoints
+     */
+
+    /**
+     * Sends credentials to the native layer
+     * @param {{username: string, password: string}} credentials
+     */
+    storeCredentials (credentials) {
+        return wkSend('pmHandlerStoreCredentials', credentials)
+    }
+
+    /**
+     * Gets the init data from the device
+     * @returns {APIResponse<PMData>}
+     */
+    async getAutofillInitData () {
+        const response = await wkSendAndWait('pmHandlerGetAutofillInitData')
+        this.storeLocalData(response.success)
+        return response
+    }
+
+    /**
+     * Gets credentials ready for autofill
+     * @param {Number} id - the credential id
+     * @returns {APIResponse<CredentialsObject>}
+     */
+    getAutofillCredentials (id) {
+        return wkSendAndWait('pmHandlerGetAutofillCredentials', { id })
+    }
+
+    /**
+     * Opens the native UI for managing passwords
+     */
+    openManagePasswords () {
+        return wkSend('pmHandlerOpenManagePasswords')
+    }
+
+    /**
+     * Opens the native UI for managing identities
+     */
+    openManageIdentities () {
+        return wkSend('pmHandlerOpenManageIdentities')
+    }
+
+    /**
+     * Opens the native UI for managing credit cards
+     */
+    openManageCreditCards () {
+        return wkSend('pmHandlerOpenManageCreditCards')
+    }
+
+    /**
+     * Gets a single identity obj once the user requests it
+     * @param {Number} id
+     * @returns {Promise<{success: IdentityObject | undefined}>}
+     */
+    getAutofillIdentity (id) {
+        const identity = this.getLocalIdentities().find(({id: identityId}) => `${identityId}` === `${id}`)
+        return Promise.resolve({success: identity})
+    }
+
+    /**
+     * Gets a single complete credit card obj once the user requests it
+     * @param {Number} id
+     * @returns {APIResponse<CreditCardObject>}
+     */
+    getAutofillCreditCard (id) {
+        return wkSendAndWait('pmHandlerGetCreditCard', { id })
+    }
+
+    async getAlias () {
+        const {alias} = await wkSendAndWait(
+            'emailHandlerGetAlias',
+            {
+                requiresUserPermission: !isApp,
+                shouldConsumeAliasIfProvided: !isApp
+            }
+        )
+        return formatDuckAddress(alias)
+    }
+}
+
+module.exports = AppleDeviceInterface
